Integration API routes
